Rename to validate.rs
