+++ contracts/tgrade-validator-voting/src/lib.rs
@@ -3,6 +3,6 @@ pub mod error;
 pub mod msg;
 #[cfg(test)]
 mod multitest;
-pub mod state;
+pub mod validate;
 
 use error::ContractError;
