+++ contracts/tgrade-validator-voting/src/validate.rs
